refactor(Hackathon):Interfaces to separate page
